# docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: WhatsApp Multiagente Platform API
  version: 1.0.0
  description: >
    API para plataforma web multiagente de WhatsApp con RBAC, 2FA, dashboard,
    backups, reportes, conectores MCP y chat de IA multi-proveedor.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3001
    description: Local
...
paths:
...
  /api/v1/ai/chat:
    post:
      tags: [AI]
      summary: Chat IA sin streaming
      operationId: aiChat
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIChatRequest"
      responses:
        "200":
          description: Respuesta de modelo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIChatResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/ai/chat/stream:
    post:
      tags: [AI]
      summary: Chat IA con streaming SSE
      operationId: aiChatStream
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIChatRequest"
      responses:
        "200":
          description: Flujo SSE de tokens
          content:
            text/event-stream:
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/ai/providers:
    get:
      tags: [AI]
      summary: Listar proveedores IA configurados
      operationId: listAIProviders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Lista de proveedores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AIProviderInfo"
  /api/v1/ai/providers/test:
    post:
      tags: [AI]
      summary: Probar conectividad de proveedor IA
      operationId: testAIProvider
      description: Solo ADMIN_TECH.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, model]
              properties:
                provider:
                  type: string
                model:
                  type: string
      responses:
        "200":
          description: Resultado prueba
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskAcceptedResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Forbidden:
      description: Sin permisos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    TaskAcceptedResponse:
      type: object
      required: [taskId, status]
      properties:
        taskId:
          type: string
        status:
          type: string
          enum: [QUEUED, PROCESSING]
    AIChatRequest:
      type: object
      required: [provider, model, prompt]
      properties:
        provider:
          type: string
          enum: [openai, gemini, claude, grok, ollama, llama_cpp]
        model:
          type: string
        prompt:
          type: string
        agentScopeId:
          type: string
          format: uuid
          nullable: true
        useTools:
          type: boolean
          default: false
        temperature:
          type: number
          format: float
          minimum: 0
          maximum: 2
    AIChatResponse:
      type: object
      required: [sessionId, provider, model, output]
      properties:
        sessionId:
          type: string
          format: uuid
        provider:
          type: string
        model:
          type: string
        output:
          type: string
        usage:
          type: object
          properties:
            inputTokens:
              type: integer
            outputTokens:
              type: integer
            estimatedCost:
              type: number
              format: float
    AIProviderInfo:
      type: object
      required: [name, enabled]
      properties:
        name:
          type: string
        enabled:
          type: boolean
        defaultModel:
          type: string
          nullable: true
        health:
          type: string
          enum: [UP, DOWN, DEGRADED]
...